chore(jenkins): convert ParameterDefinition to Java
